1.将推拉流模块集成demo工程 2.禁止activity旋转
