--- demo/libmetapush_android/libmetapush/.gitignore
@@ -1,1 +1,0 @@
-/build
