--- demo/libyangplayer_android/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Fri Dec 16 08:44:29 CST 2022
-distributionBase=GRADLE_USER_HOME
-distributionUrl=https\://services.gradle.org/distributions/gradle-7.4-bin.zip
-distributionPath=wrapper/dists
-zipStorePath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
