--- demo/libyangplayer_android/libmetaplayer/.gitignore
@@ -1,1 +1,0 @@
-/build
