+++ demo/metaplayer6_android/libmetaplayer/.gitignore
@@ -1,0 +1,3 @@
+.DS_Store
+/build
+/.cxx
