+++ demo/metaplayer6_android/libmetaplayer/build.gradle
@@ -1,21 +1,24 @@
 plugins {
-    id 'com.android.application'
+    id 'com.android.library'
 }
 
 android {
+    namespace 'com.metartc'
     compileSdk 29
 
     defaultConfig {
-        applicationId "com.metartc.player"
         minSdk 28
         targetSdk 29
-        versionCode 1
-        versionName "1.0"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-
+        consumerProguardFiles "consumer-rules.pro"
+        externalNativeBuild {
+            cmake {
+                cppFlags ''
+                abiFilters "arm64-v8a","x86_64"
+            }
+        }
     }
-
 
     buildTypes {
         release {
@@ -23,28 +26,23 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
     }
+    externalNativeBuild {
+        cmake {
+            path "src/main/cpp/CMakeLists.txt"
+            version "3.18.1"
+        }
+    }
     compileOptions {
         sourceCompatibility JavaVersion.VERSION_1_8
         targetCompatibility JavaVersion.VERSION_1_8
     }
-
-    buildFeatures {
-        viewBinding true
-    }
 }
-//repositories {
-  //  flatDir {
-    //    dir 'libs'
-   // }
-//}
+
 dependencies {
 
     implementation 'androidx.appcompat:appcompat:1.3.0'
     implementation 'com.google.android.material:material:1.4.0'
-    implementation 'androidx.constraintlayout:constraintlayout:2.0.4'
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
-    //implementation (fileTree("libs"))
-    implementation project(path: ':libmetaplayer')
 }
