+++ demo/metaplayer6_android/libmetaplayer/proguard-rules.pro
