+++ demo/metaplayer6_android/libmetaplayer/src/androidTest/java/com/metartc/ExampleInstrumentedTest.java
