+++ demo/metaplayer6_android/libmetaplayer/src/main/AndroidManifest.xml
