+++ demo/metaplayer6_android/settings.gradle
@@ -14,3 +14,4 @@ dependencyResolutionManagement {
 }
 rootProject.name = "metaplayer6"
 include ':app'
+include ':libmetaplayer'
