+++ demo/metapushstream6_android/app/build.gradle
@@ -7,7 +7,7 @@ android {
 
     defaultConfig {
         applicationId "com.metartc.push"
-        minSdk 29
+        minSdk 28
         targetSdk 29
         versionCode 1
         versionName "1.0"
@@ -40,5 +40,6 @@ dependencies {
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
-    implementation (fileTree("libs"))
+    //implementation (fileTree("libs"))
+    implementation project(path: ':libmetapush')
 }
