+++ demo/metapushstream6_android/libmetapush/.gitignore
@@ -1,0 +1,3 @@
+.DS_Store
+/build
+.cxx
