+++ demo/metapushstream6_android/libmetapush/proguard-rules.pro
