+++ demo/metapushstream6_android/libmetapush/src/androidTest/java/com/metartc/push/ExampleInstrumentedTest.java
