+++ demo/metapushstream6_android/libmetapush/src/main/AndroidManifest.xml
