+++ demo/metapushstream6_android/libmetapush/src/main/cpp/libmetapush.cpp
