+++ demo/metapushstream6_android/settings.gradle
@@ -14,3 +14,4 @@ dependencyResolutionManagement {
 }
 rootProject.name = "metapushstream6"
 include ':app'
+include ':libmetapush'
